# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Viz API
  version: 0.1.0
  description: OpenAPI specification for the Viz API (chi-based)
servers:
  - url: http://{host}:{port}
    description: Local API
    variables:
      host:
        default: localhost
        description: Hostname or IP for the API server
      port:
        default: "7770"
        description: Port for the API server
  - url: /api
    description: Production API
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /admin/system/stats:
    get:
      summary: Get system statistics (uptime, memory)
      operationId: getSystemStats
      security:
        - BearerAuth: [admin:read]
        - CookieAuth: []
      responses:
        "200":
          description: System stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/db/stats:
    get:
      summary: Get database statistics (counts)
      operationId: getDatabaseStats
      security:
        - BearerAuth: [admin:read]
        - CookieAuth: []
      responses:
        "200":
          description: Database stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseStatsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/cache/status:
    get:
      summary: Get cache status
      operationId: getCacheStatus
      security:
        - BearerAuth: [admin:read]
        - CookieAuth: []
      responses:
        "200":
          description: Cache status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheStatusResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/cache:
    delete:
      summary: Clear the image cache
      operationId: clearImageCache
      security:
        - BearerAuth: [admin:write]
        - CookieAuth: []
      responses:
        "200":
          description: Cache cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      summary: List all users
      operationId: listUsers
      security:
        - BearerAuth: [admin:read]
        - CookieAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new user (admin)
      operationId: adminCreateUser
      security:
        - BearerAuth: [admin:write]
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. The key should be provided in the Authorization header as a Bearer token. This token is a hex-encoded string of random bytes, prefixed with 'IMAG_'. Bearer format is intentionally omitted to avoid JWT confusion."
      x-scopes:
        admin:read: Read admin resources
        admin:write: Write admin resources
        api_keys:create: Create API keys
        api_keys:list: List API keys
        api_keys:get: Get API key details
        api_keys:revoke: Revoke API keys
        api_keys:rotate: Rotate API keys
        api_keys:delete: Delete API keys
        auth:refresh: Refresh authentication tokens
        auth:revoke: Revoke authentication tokens
        collections:create: Create collections
        collections:read: Read collections
        collections:update: Update collections
        collections:delete: Delete collections
        collections:share: Share collections
        images:read: Read images
        images:upload: Upload images
        images:update: Update images
        images:delete: Delete images
        images:download: Download images
        jobs:read: Read jobs
        jobs:create: Create jobs
        jobs:update: Update jobs
        jobs:delete: Delete jobs
        users:create: Create users
        users:read: Read users
        users:update: Update users
        users:delete: Delete users
        user_settings:read: Read user settings
        user_settings:update: Update user settings
        downloads:create: Create downloads
        events:read: Read events
        "*": Grants all permissions (superuser)

    CookieAuth:
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message
      required: [message]
...
    AdminUserCreate:
      type: object
      properties:
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
        role:
          type: string
          enum: [user, admin, superadmin, guest]
          default: user
          description: User role
      required: [name, email, password]
    SystemStatsResponse:
      type: object
      properties:
        uptime_seconds:
          type: integer
          format: int64
          description: System uptime in seconds
        num_goroutine:
          type: integer
          description: Number of running goroutines
        alloc_memory:
          type: integer
          format: int64
          description: "Bytes of allocated heap objects"
        sys_memory:
          type: integer
          format: int64
          description: "Total bytes of memory obtained from the OS"
        storage_used_bytes:
          type: integer
          format: int64
          description: "Total size of files in the base directory"
        storage_path:
          type: string
          description: Path to storage directory
        total_system_space_bytes:
          type: integer
          format: int64
          description: "Total disk space on the system"
        total_available_space_bytes:
          type: integer
          format: int64
          description: "Total available disk space on the system"
      required:
        [
          uptime_seconds,
          num_goroutine,
          alloc_memory,
          sys_memory,
          storage_used_bytes,
          storage_path,
          total_system_space_bytes,
          total_available_space_bytes,
        ]
    DatabaseStatsResponse:
      type: object
      properties:
        user_count:
          type: integer
          format: int64
          description: Total number of users
        image_count:
          type: integer
          format: int64
          description: Total number of images
        collection_count:
          type: integer
          format: int64
          description: Total number of collections
        db_size_bytes:
          type: integer
          format: int64
          description: "Size of the database in bytes (Postgres only)"
        active_connections:
          type: integer
          format: int64
          description: "Number of active connections (Postgres only)"
      required: [user_count, image_count, collection_count]
...
    CacheStatusResponse:
      type: object
      properties:
        size:
          type: integer
          description: Current size of the cache in bytes
        items:
          type: integer
          description: Number of items currently in the cache
        hits:
          type: integer
          description: Number of cache hits
        misses:
          type: integer
          description: Number of cache misses
        hit_ratio:
          type: number
          format: float
          description: Cache hit ratio
      required: [size, items, hits, misses, hit_ratio]
...
    User:
      x-entity: true
      type: object
      properties:
        uid: { type: string, description: User UID }
        first_name: { type: string, description: First name }
        last_name: { type: string, description: Last name }
        username: { type: string, description: Username }
        email: { type: string, description: Email }
        role:
          type: string
          enum: [user, admin, superadmin, guest]
          description: User role
        created_at:
          { type: string, format: date-time, description: Creation time }
        updated_at:
          { type: string, format: date-time, description: Update time }
      required:
        [
          uid,
          first_name,
          last_name,
          username,
          email,
          role,
          created_at,
          updated_at,
        ]
...